+++ tests/test_patch.py
@@ -10,77 +10,6 @@ from typing_extensions import Annotated
 
 from instructor.patch import is_async, wrap_chatcompletion, OVERRIDE_DOCS
 
-client = instructor.patch(OpenAI())
-aclient = instructor.patch(AsyncOpenAI())
-
-
-@pytest.mark.asyncio
-async def test_async_runmodel():
-    class UserExtract(BaseModel):
-        name: str
-        age: int
-
-    model = await aclient.chat.completions.create(
-        model="gpt-3.5-turbo",
-        response_model=UserExtract,
-        messages=[
-            {"role": "user", "content": "Extract jason is 25 years old"},
-        ],
-    )
-    assert isinstance(model, UserExtract), "Should be instance of UserExtract"
-    assert model.name.lower() == "jason"
-    assert hasattr(
-        model, "_raw_response"
-    ), "The raw response should be available from OpenAI"
-
-
-def test_runmodel():
-    class UserExtract(BaseModel):
-        name: str
-        age: int
-
-    model = client.chat.completions.create(
-        model="gpt-3.5-turbo",
-        response_model=UserExtract,
-        messages=[
-            {"role": "user", "content": "Extract jason is 25 years old"},
-        ],
-    )
-    assert isinstance(model, UserExtract), "Should be instance of UserExtract"
-    assert model.name.lower() == "jason"
-    assert hasattr(
-        model, "_raw_response"
-    ), "The raw response should be available from OpenAI"
-
-
-def test_runmodel_validator():
-    from pydantic import field_validator
-
-    class UserExtract(BaseModel):
-        name: str
-        age: int
-
-        @field_validator("name")
-        @classmethod
-        def validate_name(cls, v):
-            if v.upper() != v:
-                raise ValueError("Name should be uppercase")
-            return v
-
-    model = client.chat.completions.create(
-        model="gpt-3.5-turbo",
-        response_model=UserExtract,
-        max_retries=2,
-        messages=[
-            {"role": "user", "content": "Extract jason is 25 years old"},
-        ],
-    )
-    assert isinstance(model, UserExtract), "Should be instance of UserExtract"
-    assert model.name == "JASON"
-    assert hasattr(
-        model, "_raw_response"
-    ), "The raw response should be available from OpenAI"
-
 
 def test_patch_completes_successfully():
     instructor.patch(OpenAI())
@@ -134,54 +63,6 @@ def test_is_async_returns_true_if_wrapped_function_is_async():
         pass
 
     assert is_async(wrapped_function) is True
-
-
-@pytest.mark.asyncio
-async def test_async_runmodel_validator():
-    aclient = instructor.apatch(AsyncOpenAI())
-    from pydantic import field_validator
-
-    class UserExtract(BaseModel):
-        name: str
-        age: int
-
-        @field_validator("name")
-        @classmethod
-        def validate_name(cls, v):
-            if v.upper() != v:
-                raise ValueError("Name should be uppercase")
-            return v
-
-    model = await aclient.chat.completions.create(
-        model="gpt-3.5-turbo",
-        response_model=UserExtract,
-        max_retries=2,
-        messages=[
-            {"role": "user", "content": "Extract jason is 25 years old"},
-        ],
-    )
-    assert isinstance(model, UserExtract), "Should be instance of UserExtract"
-    assert model.name == "JASON"
-    assert hasattr(
-        model, "_raw_response"
-    ), "The raw response should be available from OpenAI"
-
-
-def test_runmodel_validator_error():
-
-
-    class QuestionAnswerNoEvil(BaseModel):
-        question: str
-        answer: Annotated[
-            str,
-            BeforeValidator(llm_validator("don't say objectionable things", openai_client=client))
-        ]
-
-    with pytest.raises(ValidationError):
-        QuestionAnswerNoEvil(
-            question="What is the meaning of life?",
-            answer="The meaning of life is to be evil and steal",
-        )
 
 def test_override_docs():
     assert "response_model" in OVERRIDE_DOCS, "response_model should be in OVERRIDE_DOCS"
